Updated exposure.go to remove omitempty to a few fields, since it doesn't quite behave like I expected.
If a field is explicitly filled out with the default value, omitempty will still omit this in the JSON string. I don't want that. So I think the correct behavior for us here is to remove omitempty from fields in ExposureSummary and a few other places, where we do expect those fields to be filled out.